separate math into own file
